+++ src/main/kotlin/quickfix/dto/chat/MessageResponseDTO.kt
@@ -1,4 +1,4 @@
-package quickfix.dto.message
+package quickfix.dto.chat
 
 data class MessageResponseDTO(
     val message: String,
